+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/project-access
 
+## 1.13.8
+
+### Patch Changes
+
+-   @sap-ux/ui5-config@0.19.5
+
 ## 1.13.7
 
 ### Patch Changes